feat: released two endpoint (irina task) in admin controller) and some refactor
