+++ main-service/src/main/java/ru/practicum/explorewithme/categories/controller/AdminCategoriesController.java
@@ -5,13 +5,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import ru.practicum.explorewithme.categories.dto.CategoryDto;
 import ru.practicum.explorewithme.categories.dto.NewCategoryDto;
 import ru.practicum.explorewithme.categories.service.CategoryService;
@@ -38,7 +32,7 @@ public class AdminCategoriesController {
 
     @PatchMapping("/{categoryId}")
     public ResponseEntity<CategoryDto> update(@PathVariable Long categoryId,
-                                       @RequestBody @Valid CategoryDto categoryDto) {
+                                              @RequestBody @Valid CategoryDto categoryDto) {
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(categoryService.updateCategory(categoryId, categoryDto));
