+++ main-service/src/main/java/ru/practicum/explorewithme/categories/dto/NewCategoryDto.java
@@ -2,13 +2,7 @@ package ru.practicum.explorewithme.categories.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
 @Getter
 @Setter
