+++ main-service/src/main/java/ru/practicum/explorewithme/comments/controller/AdminCommentsController.java
@@ -6,6 +6,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.explorewithme.comments.CommentService;
 import ru.practicum.explorewithme.comments.dto.CommentDto;
+import ru.practicum.explorewithme.comments.model.CommentStatus;
 
 @Slf4j
 @RestController
@@ -23,20 +24,20 @@ public class AdminCommentsController {
 
     @GetMapping("/{commentId}")
     @ResponseStatus(HttpStatus.OK)
-    public CommentDto findCommentById(@PathVariable Long commentId) {
+    public CommentDto findCommentById(@PathVariable long commentId) {
         return null;
     }
 
     @DeleteMapping("/{commentId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteComment(@PathVariable Long commentId) {
+    public void deleteComment(@PathVariable long commentId) {
         commentService.deleteComment(commentId);
     }
 
     @PatchMapping("/{commentId}")
     @ResponseStatus(HttpStatus.OK)
-    public CommentDto updateComment(@PathVariable Long commentId, @RequestBody CommentDto commentDto) {
-        return null;
+    public CommentDto patchCommentStatus(@PathVariable long commentId,
+                                         @RequestParam(defaultValue = "APPROVED") CommentStatus status) {
+        return commentService.patchCommentStatus(commentId, status);
     }
-
 }
