--- main-service/src/main/java/ru/practicum/explorewithme/comments/controller/PublicCommentsController.java
@@ -1,4 +1,0 @@
-package ru.practicum.explorewithme.comments.controller;
-
-public class PublicCommentsController {
-}
