--- main-service/src/main/java/ru/practicum/explorewithme/comments/dto/PrivateCommentParams.java
@@ -1,12 +1,0 @@
-package ru.practicum.explorewithme.comments.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class PrivateCommentParams {
-    private Long userId;
-    private Long eventId;
-    private NewCommentDto newCommentDto;
-}
