+++ main-service/src/main/java/ru/practicum/explorewithme/compilations/dto/NewCompilationDto.java
@@ -3,6 +3,7 @@ package ru.practicum.explorewithme.compilations.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.*;
+
 import java.util.Set;
 
 @Getter
