+++ main-service/src/main/java/ru/practicum/explorewithme/compilations/service/CompilationServiceImpl.java
@@ -14,8 +14,10 @@ import ru.practicum.explorewithme.events.model.Event;
 import ru.practicum.explorewithme.events.repository.EventRepository;
 import ru.practicum.explorewithme.exception.NotFoundException;
 
-
-import java.util.*;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
