+++ main-service/src/main/java/ru/practicum/explorewithme/events/controller/AdminEventsController.java
@@ -3,7 +3,6 @@ package ru.practicum.explorewithme.events.controller;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -11,7 +10,7 @@ import ru.practicum.explorewithme.events.dto.AdminEventParams;
 import ru.practicum.explorewithme.events.dto.EventDto;
 import ru.practicum.explorewithme.events.dto.NewEventDto;
 import ru.practicum.explorewithme.events.dto.RequestMethod;
-import ru.practicum.explorewithme.events.enumiration.EventState;
+import ru.practicum.explorewithme.events.enumeration.EventState;
 import ru.practicum.explorewithme.events.service.EventService;
 
 import java.time.LocalDateTime;
@@ -20,7 +19,6 @@ import java.util.List;
 @RestController
 @RequestMapping(path = "/admin/events")
 @RequiredArgsConstructor
-@Slf4j
 public class AdminEventsController {
     private final EventService eventService;
