+++ main-service/src/main/java/ru/practicum/explorewithme/events/dto/NewEventDto.java
@@ -11,7 +11,7 @@ import lombok.Data;
 import ru.practicum.explorewithme.config.CustomLocalDateTimeDeserializer;
 import ru.practicum.explorewithme.config.CustomLocalDateTimeSerializer;
 import ru.practicum.explorewithme.constraint.EventStartDateTime;
-import ru.practicum.explorewithme.events.enumiration.EventStateAction;
+import ru.practicum.explorewithme.events.enumeration.EventStateAction;
 
 import java.time.LocalDateTime;
