+++ main-service/src/main/java/ru/practicum/explorewithme/events/dto/UserEventParams.java
@@ -16,7 +16,7 @@ public class UserEventParams {
     private Boolean onlyAvailable;
     private LocalDateTime rangeStart;
     private LocalDateTime rangeEnd;
-    private Optional<EventSortEnum> sort;
+    private Optional<EventSortEnum> sort; // todo: не рекомендуется так делать
     private Integer from;
     private Integer size;
