+++ main-service/src/main/java/ru/practicum/explorewithme/events/enumeration/EventStateAction.java
@@ -1,4 +1,4 @@
-package ru.practicum.explorewithme.events.enumiration;
+package ru.practicum.explorewithme.events.enumeration;
 
 public enum EventStateAction {
     PUBLISH_EVENT,
