+++ main-service/src/main/java/ru/practicum/explorewithme/exception/DataAlreadyExistException.java
@@ -12,12 +12,4 @@ public class DataAlreadyExistException extends RuntimeException {
     public DataAlreadyExistException(String message, Object... args) {
         super(MessageFormat.format(message, args));
     }
-
-    public static Supplier<DataAlreadyExistException> dataAlreadyExistException(String message, Object... args) {
-        return () -> new DataAlreadyExistException(message, args);
-    }
-
-    public static Supplier<DataAlreadyExistException> dataAlreadyExistException(String message) {
-        return () -> new DataAlreadyExistException(message);
-    }
 }
