+++ main-service/src/main/java/ru/practicum/explorewithme/exception/ForbiddenException.java
@@ -7,8 +7,4 @@ public class ForbiddenException extends RuntimeException {
     public ForbiddenException(String message) {
         super(message);
     }
-
-    public static Supplier<ForbiddenException> forbiddenException(String message) {
-        return () -> new ForbiddenException(message);
-    }
 }
