--- main-service/src/main/java/ru/practicum/explorewithme/users/mapper/RequestMapper.java
@@ -1,8 +1,0 @@
-package ru.practicum.explorewithme.users.mapper;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface RequestMapper {
-
-}
