+++ main-service/src/main/java/ru/practicum/explorewithme/users/mapper/UserMapper.java
@@ -5,7 +5,6 @@ import ru.practicum.explorewithme.users.dto.NewUserRequest;
 import ru.practicum.explorewithme.users.dto.UserDto;
 import ru.practicum.explorewithme.users.model.User;
 
-
 @Mapper(componentModel = "spring")
 public interface UserMapper {
